Log out button implemented as a component for reuseability
